// src/redux/editEntityPageContent/editEntityPageContent.types.ts
export enum EditEntityPageContentActions {
  // Header
  UpdateHeaderContent = 'ixo/EditEntityPageContent/UPDATE_HEADER',
  UploadHeaderContentImage = 'ixo/EditEntityPageContent/UPLOAD_HEADER_IMAGE',
  UploadHeaderContentImagePending = 'ixo/EditEntityPageContent/UPLOAD_HEADER_IMAGE_PENDING',
  UploadHeaderContentImageSuccess = 'ixo/EditEntityPageContent/UPLOAD_HEADER_IMAGE_FULFILLED',
  UploadHeaderContentImageFailure = 'ixo/EditEntityPageContent/UPLOAD_HEADER_IMAGE_REJECTED',
  UploadHeaderContentLogo = 'ixo/EditEntityPageContent/UPLOAD_HEADER_LOGO',
  UploadHeaderContentLogoPending = 'ixo/EditEntityPageContent/UPLOAD_HEADER_LOGO_PENDING',
  UploadHeaderContentLogoSuccess = 'ixo/EditEntityPageContent/UPLOAD_HEADER_LOGO_FULFILLED',
  UploadHeaderContentLogoFailure = 'ixo/EditEntityPageContent/UPLOAD_HEADER_LOGO_REJECTED',
  // Body
  AddBodySection = 'ixo/EditEntityPageContent/ADD_BODY_SECTION',
  RemoveBodySection = 'ixo/EditEntityPageContent/REMOVE_BODY_SECTION',
  UpdateBodyContent = 'ixo/EditEntityPageContent/UPDATE_BODY',
  UploadBodyContentImage = 'ixo/EditEntityPageContent/UPLOAD_BODY_IMAGE',
  UploadBodyContentImagePending = 'ixo/EditEntityPageContent/UPLOAD_BODY_IMAGE_PENDING',
  UploadBodyContentImageSuccess = 'ixo/EditEntityPageContent/UPLOAD_BODY_IMAGE_FULFILLED',
  UploadBodyContentImageFailure = 'ixo/EditEntityPageContent/UPLOAD_BODY_IMAGE_REJECTED',
  // Image
  AddImageSection = 'ixo/EditEntityPageContent/ADD_IMAGE_SECTION',
  RemoveImageSection = 'ixo/EditEntityPageContent/REMOVE_IMAGE_SECTION',
  UpdateImageContent = 'ixo/EditEntityPageContent/UPDATE_IMAGE',
  UploadImageContentImage = 'ixo/EditEntityPageContent/UPLOAD_IMAGE_IMAGE',
  UploadImageContentImagePending = 'ixo/EditEntityPageContent/UPLOAD_IMAGE_IMAGE_PENDING',
  UploadImageContentImageSuccess = 'ixo/EditEntityPageContent/UPLOAD_IMAGE_IMAGE_FULFILLED',
  UploadImageContentImageFailure = 'ixo/EditEntityPageContent/UPLOAD_IMAGE_IMAGE_REJECTED',
  // Video
  AddVideoSection = 'ixo/EditEntityPageContent/ADD_VIDEO_SECTION',
  RemoveVideoSection = 'ixo/EditEntityPageContent/REMOVE_VIDEO_SECTION',
  UpdateVideoContent = 'ixo/EditEntityPageContent/UPDATE_VIDEO',
  UploadVideoContentVideo = 'ixo/EditEntityPageContent/UPLOAD_VIDEO_VIDEO',
  UploadVideoContentVideoPending = 'ixo/EditEntityPageContent/UPLOAD_VIDEO_VIDEO_PENDING',
  UploadVideoContentVideoSuccess = 'ixo/EditEntityPageContent/UPLOAD_VIDEO_VIDEO_FULFILLED',
  UploadVideoContentVideoFailure = 'ixo/EditEntityPageContent/UPLOAD_VIDEO_VIDEO_REJECTED',
  // Profile
  AddProfileSection = 'ixo/EditEntityPageContent/ADD_PROFILE_SECTION',
  RemoveProfileSection = 'ixo/EditEntityPageContent/REMOVE_PROFILE_SECTION',
  UpdateProfileContent = 'ixo/EditEntityPageContent/UPDATE_PROFILE',
  UploadProfileContentImage = 'ixo/EditEntityPageContent/UPLOAD_PROFILE_IMAGE',
  UploadProfileContentImagePending = 'ixo/EditEntityPageContent/UPLOAD_PROFILE_IMAGE_PENDING',
  UploadProfileContentImageSuccess = 'ixo/EditEntityPageContent/UPLOAD_PROFILE_IMAGE_FULFILLED',
  UploadProfileContentImageFailure = 'ixo/EditEntityPageContent/UPLOAD_PROFILE_IMAGE_REJECTED',
  // Social
  UpdateSocialContent = 'ixo/EditEntityPageContent/UPDATE_SOCIAL',
  // Embedded
  AddEmbeddedSection = 'ixo/EditEntityPageContent/ADD_EMBEDDED_SECTION',
  RemoveEmbeddedSection = 'ixo/EditEntityPageContent/REMOVE_EMBEDDED_SECTION',
  UpdateEmbeddedContent = 'ixo/EditEntityPageContent/UPDATE_EMBEDDED',
  // Validation
  Validated = 'ixo/EditEntityPageContent/SET_VALIDATED',
  ValidationError = 'ixo/EditEntityPageContent/VALIDATION_ERROR',

  // Import Page Contnet
  ImportEntityPageContent = 'ixo/EditEntityPageContent/IMPORT_ENTITY_PAGE_CONTENT',

  // Ordering Page Content
  OrderEntityPageContent = 'ixo/EditEntityPageContent/ORDER_ENTITY_PAGE_CONTENT',
}